+++ src/main/java/com/example/my_gradle_spring_app/repository/ProblemRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 
 @Repository
 public interface ProblemRepository extends JpaRepository<Problem, Long> {
-    List<Problem> findByContestId(Integer contestId);
+    List<Problem> findByContestId(Long contestId);
 }
